+++ src/main/java/chumbanotz/mutantbeasts/client/model/ScalableModelRenderer.java
@@ -4,8 +4,9 @@ import net.minecraft.client.model.ModelBase;
 import net.minecraft.client.model.ModelRenderer;
 import net.minecraft.client.renderer.GlStateManager;
 
-public class ScalableModelRenderer extends ModelRenderer {
-    private float scale = 1.0F;
+public class ScalableModelRenderer
+extends ModelRenderer {
+    private float scale = 1.0f;
 
     public ScalableModelRenderer(ModelBase model, int texOffX, int texOffY) {
         super(model, texOffX, texOffY);
@@ -18,7 +19,7 @@ public class ScalableModelRenderer extends ModelRenderer {
     public void render(float scale) {
         if (!this.isHidden && this.showModel) {
             GlStateManager.pushMatrix();
-            GlStateManager.scale(this.scale, this.scale, this.scale);
+            GlStateManager.scale((float)this.scale, (float)this.scale, (float)this.scale);
             super.render(scale);
             GlStateManager.popMatrix();
         }
